make app refresh when there's new data in database
Changes to username, phone number, password, agenda, announcement, and favorite app will now be visible in the app in 'real time', but banner updates and the ability to switch language and notifications are still wip. When a user presses the language and notification switch, an alert dialog will pop up saying that this feature is still under development.